Work using 'clone' with namespaces.
It will likely be more convinient to use unshare with fork.  This will
mess up the parent process, but that's not really a big deal, since we
won't be able to track it anyways.

It would be nice though, if the process stayed connected with
stdin/stdout.  This should be possible if we unshare() and execve() but
then we won't get a predicable process id?

There has to be a solution for this?!