Spawn application process
This almost seems to work, we do get a process id back, but we fault in
'child_main' when using 'println!'.

I suspect, that Rust just isn't able to deal with 'fork()' or 'clone()'.
Instead, we should implement that function in C or C++.  That should
make many things easier.  Just put that in 'util::'.