We appear to loose 'root' privileges on 'clone()'
My understanding is, that 'clone()' creates a new user with all
capabilites when 'CLONE_NEWUSER' is specified.

Somehow, this user is not considered root (UID=0) in the new user
namespace though.  If we trust 'user_namespaces(7)', then we loose all
capabilities on 'execve' unless we have 'UID=0'.

Maybe I should look at the implementation of 'CLONE_NEWUSER'.